third_party/blink/web_tests/fast/forms/range/range-value-rounding-expected.txt

80009 Introduce Decimal arithmetic to fix rounding errors in number/range input types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS createSample("0.1", "0.1", "1e-0").value is "0.1"
PASS createSample("0.1", "0.1", "1e-0").valueAsNumber is 0.1
PASS createSample("0.1", "0.2", "1e-0").value is "0.2"
PASS createSample("0.1", "0.2", "1e-0").valueAsNumber is 0.2
PASS createSample("0.1", "0.3", "1e-0").value is "0.3"
PASS createSample("0.1", "0.3", "1e-0").valueAsNumber is 0.3
PASS createSample("0.1", "0.4", "1e-0").value is "0.4"
PASS createSample("0.1", "0.4", "1e-0").valueAsNumber is 0.4
PASS createSample("0.1", "0.5", "1e-0").value is "0.5"
PASS createSample("0.1", "0.5", "1e-0").valueAsNumber is 0.5
PASS createSample("0.1", "0.6", "1e-0").value is "0.6"
PASS createSample("0.1", "0.6", "1e-0").valueAsNumber is 0.6
PASS createSample("0.1", "0.7", "1e-0").value is "0.7"
PASS createSample("0.1", "0.7", "1e-0").valueAsNumber is 0.7
PASS createSample("0.1", "0.8", "1e-0").value is "0.8"
PASS createSample("0.1", "0.8", "1e-0").valueAsNumber is 0.8
PASS createSample("0.1", "0.9", "1e-0").value is "0.9"
PASS createSample("0.1", "0.9", "1e-0").valueAsNumber is 0.9
PASS createSample("0.01", "0.01", "1e-1").value is "0.01"
PASS createSample("0.01", "0.01", "1e-1").valueAsNumber is 0.01
PASS createSample("0.01", "0.02", "1e-1").value is "0.02"
PASS createSample("0.01", "0.02", "1e-1").valueAsNumber is 0.02
PASS createSample("0.01", "0.03", "1e-1").value is "0.03"
PASS createSample("0.01", "0.03", "1e-1").valueAsNumber is 0.03
PASS createSample("0.01", "0.04", "1e-1").value is "0.04"
PASS createSample("0.01", "0.04", "1e-1").valueAsNumber is 0.04
PASS createSample("0.01", "0.05", "1e-1").value is "0.05"
PASS createSample("0.01", "0.05", "1e-1").valueAsNumber is 0.05
PASS createSample("0.01", "0.06", "1e-1").value is "0.06"
PASS createSample("0.01", "0.06", "1e-1").valueAsNumber is 0.06
PASS createSample("0.01", "0.07", "1e-1").value is "0.07"
PASS createSample("0.01", "0.07", "1e-1").valueAsNumber is 0.07
PASS createSample("0.01", "0.08", "1e-1").value is "0.08"
PASS createSample("0.01", "0.08", "1e-1").valueAsNumber is 0.08
PASS createSample("0.01", "0.09", "1e-1").value is "0.09"
PASS createSample("0.01", "0.09", "1e-1").valueAsNumber is 0.09
PASS createSample("0.001", "0.001", "1e-2").value is "0.001"
PASS createSample("0.001", "0.001", "1e-2").valueAsNumber is 0.001
PASS createSample("0.001", "0.002", "1e-2").value is "0.002"
PASS createSample("0.001", "0.002", "1e-2").valueAsNumber is 0.002
PASS createSample("0.001", "0.003", "1e-2").value is "0.003"
PASS createSample("0.001", "0.003", "1e-2").valueAsNumber is 0.003
PASS createSample("0.001", "0.004", "1e-2").value is "0.004"
PASS createSample("0.001", "0.004", "1e-2").valueAsNumber is 0.004
PASS createSample("0.001", "0.005", "1e-2").value is "0.005"
PASS createSample("0.001", "0.005", "1e-2").valueAsNumber is 0.005
PASS createSample("0.001", "0.006", "1e-2").value is "0.006"
PASS createSample("0.001", "0.006", "1e-2").valueAsNumber is 0.006
PASS createSample("0.001", "0.007", "1e-2").value is "0.007"
PASS createSample("0.001", "0.007", "1e-2").valueAsNumber is 0.007
PASS createSample("0.001", "0.008", "1e-2").value is "0.008"
PASS createSample("0.001", "0.008", "1e-2").valueAsNumber is 0.008
PASS createSample("0.001", "0.009", "1e-2").value is "0.009"
PASS createSample("0.001", "0.009", "1e-2").valueAsNumber is 0.009
PASS createSample("0.0001", "0.0001", "1e-3").value is "0.0001"
PASS createSample("0.0001", "0.0001", "1e-3").valueAsNumber is 0.0001
PASS createSample("0.0001", "0.0002", "1e-3").value is "0.0002"
PASS createSample("0.0001", "0.0002", "1e-3").valueAsNumber is 0.0002
PASS createSample("0.0001", "0.0003", "1e-3").value is "0.0003"
PASS createSample("0.0001", "0.0003", "1e-3").valueAsNumber is 0.0003
PASS createSample("0.0001", "0.0004", "1e-3").value is "0.0004"
PASS createSample("0.0001", "0.0004", "1e-3").valueAsNumber is 0.0004
PASS createSample("0.0001", "0.0005", "1e-3").value is "0.0005"
PASS createSample("0.0001", "0.0005", "1e-3").valueAsNumber is 0.0005
PASS createSample("0.0001", "0.0006", "1e-3").value is "0.0006"
PASS createSample("0.0001", "0.0006", "1e-3").valueAsNumber is 0.0006
PASS createSample("0.0001", "0.0007", "1e-3").value is "0.0007"
PASS createSample("0.0001", "0.0007", "1e-3").valueAsNumber is 0.0007
PASS createSample("0.0001", "0.0008", "1e-3").value is "0.0008"
PASS createSample("0.0001", "0.0008", "1e-3").valueAsNumber is 0.0008
PASS createSample("0.0001", "0.0009", "1e-3").value is "0.0009"
PASS createSample("0.0001", "0.0009", "1e-3").valueAsNumber is 0.0009
PASS createSample("0.00001", "0.00001", "1e-4").value is "0.00001"
PASS createSample("0.00001", "0.00001", "1e-4").valueAsNumber is 0.00001
PASS createSample("0.00001", "0.00002", "1e-4").value is "0.00002"
PASS createSample("0.00001", "0.00002", "1e-4").valueAsNumber is 0.00002
PASS createSample("0.00001", "0.00003", "1e-4").value is "0.00003"
PASS createSample("0.00001", "0.00003", "1e-4").valueAsNumber is 0.00003
PASS createSample("0.00001", "0.00004", "1e-4").value is "0.00004"
PASS createSample("0.00001", "0.00004", "1e-4").valueAsNumber is 0.00004
PASS createSample("0.00001", "0.00005", "1e-4").value is "0.00005"
PASS createSample("0.00001", "0.00005", "1e-4").valueAsNumber is 0.00005
PASS createSample("0.00001", "0.00006", "1e-4").value is "0.00006"
PASS createSample("0.00001", "0.00006", "1e-4").valueAsNumber is 0.00006
PASS createSample("0.00001", "0.00007", "1e-4").value is "0.00007"
PASS createSample("0.00001", "0.00007", "1e-4").valueAsNumber is 0.00007
PASS createSample("0.00001", "0.00008", "1e-4").value is "0.00008"
PASS createSample("0.00001", "0.00008", "1e-4").valueAsNumber is 0.00008
PASS createSample("0.00001", "0.00009", "1e-4").value is "0.00009"
PASS createSample("0.00001", "0.00009", "1e-4").valueAsNumber is 0.00009
PASS createSample("0.000001", "0.000001", "1e-5").value is "0.000001"
PASS createSample("0.000001", "0.000001", "1e-5").valueAsNumber is 0.000001
PASS createSample("0.000001", "0.000002", "1e-5").value is "0.000002"
PASS createSample("0.000001", "0.000002", "1e-5").valueAsNumber is 0.000002
PASS createSample("0.000001", "0.000003", "1e-5").value is "0.000003"
PASS createSample("0.000001", "0.000003", "1e-5").valueAsNumber is 0.000003
PASS createSample("0.000001", "0.000004", "1e-5").value is "0.000004"
PASS createSample("0.000001", "0.000004", "1e-5").valueAsNumber is 0.000004
PASS createSample("0.000001", "0.000005", "1e-5").value is "0.000005"
PASS createSample("0.000001", "0.000005", "1e-5").valueAsNumber is 0.000005
PASS createSample("0.000001", "0.000006", "1e-5").value is "0.000006"
PASS createSample("0.000001", "0.000006", "1e-5").valueAsNumber is 0.000006
PASS createSample("0.000001", "0.000007", "1e-5").value is "0.000007"
PASS createSample("0.000001", "0.000007", "1e-5").valueAsNumber is 0.000007
PASS createSample("0.000001", "0.000008", "1e-5").value is "0.000008"
PASS createSample("0.000001", "0.000008", "1e-5").valueAsNumber is 0.000008
PASS createSample("0.000001", "0.000009", "1e-5").value is "0.000009"
PASS createSample("0.000001", "0.000009", "1e-5").valueAsNumber is 0.000009
PASS createSample("0.0000001", "0.0000001", "1e-6").value is "1e-7"
PASS createSample("0.0000001", "0.0000001", "1e-6").valueAsNumber is 0.0000001
PASS createSample("0.0000001", "0.0000002", "1e-6").value is "2e-7"
PASS createSample("0.0000001", "0.0000002", "1e-6").valueAsNumber is 0.0000002
PASS createSample("0.0000001", "0.0000003", "1e-6").value is "3e-7"
PASS createSample("0.0000001", "0.0000003", "1e-6").valueAsNumber is 0.0000003
PASS createSample("0.0000001", "0.0000004", "1e-6").value is "4e-7"
PASS createSample("0.0000001", "0.0000004", "1e-6").valueAsNumber is 0.0000004
PASS createSample("0.0000001", "0.0000005", "1e-6").value is "5e-7"
PASS createSample("0.0000001", "0.0000005", "1e-6").valueAsNumber is 0.0000005
PASS createSample("0.0000001", "0.0000006", "1e-6").value is "6e-7"
PASS createSample("0.0000001", "0.0000006", "1e-6").valueAsNumber is 0.0000006
PASS createSample("0.0000001", "0.0000007", "1e-6").value is "7e-7"
PASS createSample("0.0000001", "0.0000007", "1e-6").valueAsNumber is 0.0000007
PASS createSample("0.0000001", "0.0000008", "1e-6").value is "8e-7"
PASS createSample("0.0000001", "0.0000008", "1e-6").valueAsNumber is 0.0000008
PASS createSample("0.0000001", "0.0000009", "1e-6").value is "9e-7"
PASS createSample("0.0000001", "0.0000009", "1e-6").valueAsNumber is 0.0000009
PASS createSample("0.00000001", "0.00000001", "1e-7").value is "1e-8"
PASS createSample("0.00000001", "0.00000001", "1e-7").valueAsNumber is 0.00000001
PASS createSample("0.00000001", "0.00000002", "1e-7").value is "2e-8"
PASS createSample("0.00000001", "0.00000002", "1e-7").valueAsNumber is 0.00000002
PASS createSample("0.00000001", "0.00000003", "1e-7").value is "3e-8"
PASS createSample("0.00000001", "0.00000003", "1e-7").valueAsNumber is 0.00000003
PASS createSample("0.00000001", "0.00000004", "1e-7").value is "4e-8"
PASS createSample("0.00000001", "0.00000004", "1e-7").valueAsNumber is 0.00000004
PASS createSample("0.00000001", "0.00000005", "1e-7").value is "5e-8"
PASS createSample("0.00000001", "0.00000005", "1e-7").valueAsNumber is 0.00000005
PASS createSample("0.00000001", "0.00000006", "1e-7").value is "6e-8"
PASS createSample("0.00000001", "0.00000006", "1e-7").valueAsNumber is 0.00000006
PASS createSample("0.00000001", "0.00000007", "1e-7").value is "7e-8"
PASS createSample("0.00000001", "0.00000007", "1e-7").valueAsNumber is 0.00000007
PASS createSample("0.00000001", "0.00000008", "1e-7").value is "8e-8"
PASS createSample("0.00000001", "0.00000008", "1e-7").valueAsNumber is 0.00000008
PASS createSample("0.00000001", "0.00000009", "1e-7").value is "9e-8"
PASS createSample("0.00000001", "0.00000009", "1e-7").valueAsNumber is 0.00000009
PASS createSample("0.000000001", "0.000000001", "1e-8").value is "1e-9"
PASS createSample("0.000000001", "0.000000001", "1e-8").valueAsNumber is 0.000000001
PASS createSample("0.000000001", "0.000000002", "1e-8").value is "2e-9"
PASS createSample("0.000000001", "0.000000002", "1e-8").valueAsNumber is 0.000000002
PASS createSample("0.000000001", "0.000000003", "1e-8").value is "3e-9"
PASS createSample("0.000000001", "0.000000003", "1e-8").valueAsNumber is 0.000000003
PASS createSample("0.000000001", "0.000000004", "1e-8").value is "4e-9"
PASS createSample("0.000000001", "0.000000004", "1e-8").valueAsNumber is 0.000000004
PASS createSample("0.000000001", "0.000000005", "1e-8").value is "5e-9"
PASS createSample("0.000000001", "0.000000005", "1e-8").valueAsNumber is 0.000000005
PASS createSample("0.000000001", "0.000000006", "1e-8").value is "6e-9"
PASS createSample("0.000000001", "0.000000006", "1e-8").valueAsNumber is 0.000000006
PASS createSample("0.000000001", "0.000000007", "1e-8").value is "7e-9"
PASS createSample("0.000000001", "0.000000007", "1e-8").valueAsNumber is 0.000000007
PASS createSample("0.000000001", "0.000000008", "1e-8").value is "8e-9"
PASS createSample("0.000000001", "0.000000008", "1e-8").valueAsNumber is 0.000000008
PASS createSample("0.000000001", "0.000000009", "1e-8").value is "9e-9"
PASS createSample("0.000000001", "0.000000009", "1e-8").valueAsNumber is 0.000000009
PASS createSample("0.0000000001", "0.0000000001", "1e-9").value is "1e-10"
PASS createSample("0.0000000001", "0.0000000001", "1e-9").valueAsNumber is 0.0000000001
PASS createSample("0.0000000001", "0.0000000002", "1e-9").value is "2e-10"
PASS createSample("0.0000000001", "0.0000000002", "1e-9").valueAsNumber is 0.0000000002
PASS createSample("0.0000000001", "0.0000000003", "1e-9").value is "3e-10"
PASS createSample("0.0000000001", "0.0000000003", "1e-9").valueAsNumber is 0.0000000003
PASS createSample("0.0000000001", "0.0000000004", "1e-9").value is "4e-10"
PASS createSample("0.0000000001", "0.0000000004", "1e-9").valueAsNumber is 0.0000000004
PASS createSample("0.0000000001", "0.0000000005", "1e-9").value is "5e-10"
PASS createSample("0.0000000001", "0.0000000005", "1e-9").valueAsNumber is 0.0000000005
PASS createSample("0.0000000001", "0.0000000006", "1e-9").value is "6e-10"
PASS createSample("0.0000000001", "0.0000000006", "1e-9").valueAsNumber is 0.0000000006
PASS createSample("0.0000000001", "0.0000000007", "1e-9").value is "7e-10"
PASS createSample("0.0000000001", "0.0000000007", "1e-9").valueAsNumber is 0.0000000007
PASS createSample("0.0000000001", "0.0000000008", "1e-9").value is "8e-10"
PASS createSample("0.0000000001", "0.0000000008", "1e-9").valueAsNumber is 0.0000000008
PASS createSample("0.0000000001", "0.0000000009", "1e-9").value is "9e-10"
PASS createSample("0.0000000001", "0.0000000009", "1e-9").valueAsNumber is 0.0000000009
PASS createSample("0.00000000001", "0.00000000001", "1e-10").value is "1e-11"
PASS createSample("0.00000000001", "0.00000000001", "1e-10").valueAsNumber is 0.00000000001
PASS createSample("0.00000000001", "0.00000000002", "1e-10").value is "2e-11"
PASS createSample("0.00000000001", "0.00000000002", "1e-10").valueAsNumber is 0.00000000002
PASS createSample("0.00000000001", "0.00000000003", "1e-10").value is "3e-11"
PASS createSample("0.00000000001", "0.00000000003", "1e-10").valueAsNumber is 0.00000000003
PASS createSample("0.00000000001", "0.00000000004", "1e-10").value is "4e-11"
PASS createSample("0.00000000001", "0.00000000004", "1e-10").valueAsNumber is 0.00000000004
PASS createSample("0.00000000001", "0.00000000005", "1e-10").value is "5e-11"
PASS createSample("0.00000000001", "0.00000000005", "1e-10").valueAsNumber is 0.00000000005
PASS createSample("0.00000000001", "0.00000000006", "1e-10").value is "6e-11"
PASS createSample("0.00000000001", "0.00000000006", "1e-10").valueAsNumber is 0.00000000006
PASS createSample("0.00000000001", "0.00000000007", "1e-10").value is "7e-11"
PASS createSample("0.00000000001", "0.00000000007", "1e-10").valueAsNumber is 0.00000000007
PASS createSample("0.00000000001", "0.00000000008", "1e-10").value is "8e-11"
PASS createSample("0.00000000001", "0.00000000008", "1e-10").valueAsNumber is 0.00000000008
PASS createSample("0.00000000001", "0.00000000009", "1e-10").value is "9e-11"
PASS createSample("0.00000000001", "0.00000000009", "1e-10").valueAsNumber is 0.00000000009
PASS createSample("0.000000000001", "0.000000000001", "1e-11").value is "1e-12"
PASS createSample("0.000000000001", "0.000000000001", "1e-11").valueAsNumber is 0.000000000001
PASS createSample("0.000000000001", "0.000000000002", "1e-11").value is "2e-12"
PASS createSample("0.000000000001", "0.000000000002", "1e-11").valueAsNumber is 0.000000000002
PASS createSample("0.000000000001", "0.000000000003", "1e-11").value is "3e-12"
PASS createSample("0.000000000001", "0.000000000003", "1e-11").valueAsNumber is 0.000000000003
PASS createSample("0.000000000001", "0.000000000004", "1e-11").value is "4e-12"
PASS createSample("0.000000000001", "0.000000000004", "1e-11").valueAsNumber is 0.000000000004
PASS createSample("0.000000000001", "0.000000000005", "1e-11").value is "5e-12"
PASS createSample("0.000000000001", "0.000000000005", "1e-11").valueAsNumber is 0.000000000005
PASS createSample("0.000000000001", "0.000000000006", "1e-11").value is "6e-12"
PASS createSample("0.000000000001", "0.000000000006", "1e-11").valueAsNumber is 0.000000000006
PASS createSample("0.000000000001", "0.000000000007", "1e-11").value is "7e-12"
PASS createSample("0.000000000001", "0.000000000007", "1e-11").valueAsNumber is 0.000000000007
PASS createSample("0.000000000001", "0.000000000008", "1e-11").value is "8e-12"
PASS createSample("0.000000000001", "0.000000000008", "1e-11").valueAsNumber is 0.000000000008
PASS createSample("0.000000000001", "0.000000000009", "1e-11").value is "9e-12"
PASS createSample("0.000000000001", "0.000000000009", "1e-11").valueAsNumber is 0.000000000009
PASS createSample("0.0000000000001", "0.0000000000001", "1e-12").value is "1e-13"
PASS createSample("0.0000000000001", "0.0000000000001", "1e-12").valueAsNumber is 0.0000000000001
PASS createSample("0.0000000000001", "0.0000000000002", "1e-12").value is "2e-13"
PASS createSample("0.0000000000001", "0.0000000000002", "1e-12").valueAsNumber is 0.0000000000002
PASS createSample("0.0000000000001", "0.0000000000003", "1e-12").value is "3e-13"
PASS createSample("0.0000000000001", "0.0000000000003", "1e-12").valueAsNumber is 0.0000000000003
PASS createSample("0.0000000000001", "0.0000000000004", "1e-12").value is "4e-13"
PASS createSample("0.0000000000001", "0.0000000000004", "1e-12").valueAsNumber is 0.0000000000004
PASS createSample("0.0000000000001", "0.0000000000005", "1e-12").value is "5e-13"
PASS createSample("0.0000000000001", "0.0000000000005", "1e-12").valueAsNumber is 0.0000000000005
PASS createSample("0.0000000000001", "0.0000000000006", "1e-12").value is "6e-13"
PASS createSample("0.0000000000001", "0.0000000000006", "1e-12").valueAsNumber is 0.0000000000006
PASS createSample("0.0000000000001", "0.0000000000007", "1e-12").value is "7e-13"
PASS createSample("0.0000000000001", "0.0000000000007", "1e-12").valueAsNumber is 0.0000000000007
PASS createSample("0.0000000000001", "0.0000000000008", "1e-12").value is "8e-13"
PASS createSample("0.0000000000001", "0.0000000000008", "1e-12").valueAsNumber is 0.0000000000008
PASS createSample("0.0000000000001", "0.0000000000009", "1e-12").value is "9e-13"
PASS createSample("0.0000000000001", "0.0000000000009", "1e-12").valueAsNumber is 0.0000000000009
PASS createSample("0.00000000000001", "0.00000000000001", "1e-13").value is "1e-14"
PASS createSample("0.00000000000001", "0.00000000000001", "1e-13").valueAsNumber is 0.00000000000001
PASS createSample("0.00000000000001", "0.00000000000002", "1e-13").value is "2e-14"
PASS createSample("0.00000000000001", "0.00000000000002", "1e-13").valueAsNumber is 0.00000000000002
PASS createSample("0.00000000000001", "0.00000000000003", "1e-13").value is "3e-14"
PASS createSample("0.00000000000001", "0.00000000000003", "1e-13").valueAsNumber is 0.00000000000003
PASS createSample("0.00000000000001", "0.00000000000004", "1e-13").value is "4e-14"
PASS createSample("0.00000000000001", "0.00000000000004", "1e-13").valueAsNumber is 0.00000000000004
PASS createSample("0.00000000000001", "0.00000000000005", "1e-13").value is "5e-14"
PASS createSample("0.00000000000001", "0.00000000000005", "1e-13").valueAsNumber is 0.00000000000005
PASS createSample("0.00000000000001", "0.00000000000006", "1e-13").value is "6e-14"
PASS createSample("0.00000000000001", "0.00000000000006", "1e-13").valueAsNumber is 0.00000000000006
PASS createSample("0.00000000000001", "0.00000000000007", "1e-13").value is "7e-14"
PASS createSample("0.00000000000001", "0.00000000000007", "1e-13").valueAsNumber is 0.00000000000007
PASS createSample("0.00000000000001", "0.00000000000008", "1e-13").value is "8e-14"
PASS createSample("0.00000000000001", "0.00000000000008", "1e-13").valueAsNumber is 0.00000000000008
PASS createSample("0.00000000000001", "0.00000000000009", "1e-13").value is "9e-14"
PASS createSample("0.00000000000001", "0.00000000000009", "1e-13").valueAsNumber is 0.00000000000009
PASS createSample("0.000000000000001", "0.000000000000001", "1e-14").value is "1e-15"
PASS createSample("0.000000000000001", "0.000000000000001", "1e-14").valueAsNumber is 0.000000000000001
PASS createSample("0.000000000000001", "0.000000000000002", "1e-14").value is "2e-15"
PASS createSample("0.000000000000001", "0.000000000000002", "1e-14").valueAsNumber is 0.000000000000002
PASS createSample("0.000000000000001", "0.000000000000003", "1e-14").value is "3e-15"
PASS createSample("0.000000000000001", "0.000000000000003", "1e-14").valueAsNumber is 0.000000000000003
PASS createSample("0.000000000000001", "0.000000000000004", "1e-14").value is "4e-15"
PASS createSample("0.000000000000001", "0.000000000000004", "1e-14").valueAsNumber is 0.000000000000004
PASS createSample("0.000000000000001", "0.000000000000005", "1e-14").value is "5e-15"
PASS createSample("0.000000000000001", "0.000000000000005", "1e-14").valueAsNumber is 0.000000000000005
PASS createSample("0.000000000000001", "0.000000000000006", "1e-14").value is "6e-15"
PASS createSample("0.000000000000001", "0.000000000000006", "1e-14").valueAsNumber is 0.000000000000006
PASS createSample("0.000000000000001", "0.000000000000007", "1e-14").value is "7e-15"
PASS createSample("0.000000000000001", "0.000000000000007", "1e-14").valueAsNumber is 0.000000000000007
PASS createSample("0.000000000000001", "0.000000000000008", "1e-14").value is "8e-15"
PASS createSample("0.000000000000001", "0.000000000000008", "1e-14").valueAsNumber is 0.000000000000008
PASS createSample("0.000000000000001", "0.000000000000009", "1e-14").value is "9e-15"
PASS createSample("0.000000000000001", "0.000000000000009", "1e-14").valueAsNumber is 0.000000000000009
PASS successfullyParsed is true

TEST COMPLETE
